@@ -1,6 +1,5 @@
 import 'package:dio/adapter.dart';
 import 'package:dio/dio.dart';
-import 'package:meta/meta.dart';
 import 'package:qiniu_sdk_base/src/storage/error/error.dart';
 
 part 'protocol.dart';
@@ -18,9 +17,9 @@ class Config {
   final int retryLimit;
 
   Config({
-    HostProvider hostProvider,
-    CacheProvider cacheProvider,
-    HttpClientAdapter httpClientAdapter,
+    HostProvider? hostProvider,
+    CacheProvider? cacheProvider,
+    HttpClientAdapter? httpClientAdapter,
     this.retryLimit = 3,
   })  : hostProvider = hostProvider ?? DefaultHostProvider(),
         cacheProvider = cacheProvider ?? DefaultCacheProvider(),
