@@ -2,6 +2,7 @@ import 'dart:io';
 import 'package:qiniu_sdk_base/src/storage/methods/put/by_single/put_by_single_task.dart';
 import 'package:qiniu_sdk_base/src/storage/methods/put/put_response.dart';
 import 'package:test/test.dart';
+// ignore: import_of_legacy_library_into_null_safe
 import 'package:dotenv/dotenv.dart' show env;
 import 'package:qiniu_sdk_base/qiniu_sdk_base.dart';
 
@@ -47,15 +48,15 @@ void main() {
     final storage = Storage();
 
     final auth = Auth(
-      accessKey: env['QINIU_DART_SDK_ACCESS_KEY'],
-      secretKey: env['QINIU_DART_SDK_SECRET_KEY'],
+      accessKey: env['QINIU_DART_SDK_ACCESS_KEY']!,
+      secretKey: env['QINIU_DART_SDK_SECRET_KEY']!,
     );
 
     final token = auth.generateUploadToken(
       putPolicy: PutPolicy(
         insertOnly: 0,
         returnBody: '{"ext": \$(ext)}',
-        scope: env['QINIU_DART_SDK_TOKEN_SCOPE'],
+        scope: env['QINIU_DART_SDK_TOKEN_SCOPE']!,
         deadline: DateTime.now().millisecondsSinceEpoch + 3600,
       ),
     );
